migrate encoding/json/v2
